requirement.txt
numpy==1.20.3
opencv-python-headless==4.5.3
google-colab==1.0.0
